small fix.
git-svn-id: svn://activequant.com/svn/repositories/aq2o/trunk@553 01b8edfa-0b2f-4ad0-b428-228ee2dd3b63

+++ src/main/java/com/activequant/backtesting/BacktestConfiguration.java
@@ -31,7 +31,7 @@ public class BacktestConfiguration extends PersistentEntity {
 	}
 
 	@Property
-	public double getDate8Time6Start() {
+	public long getDate8Time6Start() {
 		return date8Time6Start;
 	}
 
@@ -40,7 +40,7 @@ public class BacktestConfiguration extends PersistentEntity {
 	}
 
 	@Property
-	public double getDate8Time6End() {
+	public long getDate8Time6End() {
 		return date8Time6End;
 	}
